Add files via upload
HTML Structure: HTML provides the structure of your web page. It contains elements that define the content and layout.
CSS Styling: CSS is used to style the HTML elements.
JavaScript Interactivity: JavaScript adds interactivity and dynamic behavior to your web page.
When linking external CSS and JavaScript files, use relative paths to specify the file location. In the examples above, "styles.css" and "script.js" are assumed to be in the same directory as the HTML file. You can use "../" to navigate up a directory if needed.